fix(stdio): processstartmode.inheritstdio incompat with inheritstdio
When using `mode: ProcessStartMode.inheritStdio` this causes the process
to hang if `inheritStdio: true`.

Also the naming here is confusing, in a future "MAJOR" version this will
resolved. But basically `inheritStdio` is our async functionality that
allows multiple processes to interleave their stdio.

Where as `ProcessStartMode.inheritStdio` is darts built-in way of directly
connecting the child process to the parent stdio streams.

This fix makes these options mutually exclusive.